+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Fixed
+- Fixed bug where tokens for the wrong app could be retrieved
+
 ## 2.0.3 - 2019-10-03
 ### Added
 - Added `craft.oauth` Twig variable